fix: fixed return value, changed fromm Integer to Long

+++ back-end/src/main/java/com/ednaldoluiz/moviedash/repository/projection/genre/GenreProjection.java
@@ -4,6 +4,6 @@ import java.io.Serializable;
 
 public record GenreProjection(
         String name,
-        Integer quantity,
-        Integer total
+        Long quantity,
+        Long total
         ) implements Serializable {}
